requirements.txt
Flask>=2.3.0
gunicorn>=21.2.0
cryptography>=41.0.7
psutil>=5.9.0
requests>=2.31.0
psycopg2-binary>=2.9.0